+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+22nd April 2021 - [AlexHedley](https://github.com/alexhedley)
+- Added Picker Wheel
+
 23rd November 2019 - [AlexHedley](https://github.com/alexhedley)
 - Using `localstorage` to remember last selected tab